requirements.txt
absl-py==1.4.0
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
altair-data-server==0.4.1
altair-saver==0.5.0
altair-viewer==0.4.0
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.2.0
audioread==3.0.1
autograd==1.6.2
autograd-gamma==0.5.0
Automat==22.10.0
Babel==2.15.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
cachetools==5.3.0
catboost==1.1.1
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==3.0.1
click==8.1.7
cloudpickle==2.2.1
colorama==0.4.6
comm==0.1.3
configparser==7.0.0
constantly==23.10.4
contourpy==1.0.7
coverage==7.6.0
coveralls==4.0.1
crispy-bootstrap4==2024.1
cryptography==42.0.8
cycler==0.11.0
dataclasses-json==0.6.6
deap==1.2.2
debugpy==1.6.7
decorator==4.4.2
defusedxml==0.7.1
distlib==0.3.6
distro==1.9.0
Django==5.0.6
django-allauth==0.63.3
django-allauth-2fa==0.11.1
django-cors-headers==4.4.0
django-crispy-forms==2.1
django-embed-video==1.4.10
django-otp==1.5.0
djangorestframework==3.15.2
docopt==0.6.2
docstring_parser==0.16
elm==0.1.1
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.2.1
factor_analyzer==0.5.1
fastjsonschema==2.19.1
filelock==3.9.0
flatbuffers==23.1.21
fonttools==4.38.0
formulaic==1.0.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.6.1
future==1.0.0
fuzzytree==0.1.4
gast==0.4.0
genson==1.3.0
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.131.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==0.4.6
google-cloud-aiplatform==1.49.0
google-cloud-bigquery==3.23.1
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.3
google-cloud-speech==2.26.0
google-cloud-storage==2.16.0
google-cloud-videointelligence==2.13.3
google-cloud-vision==3.7.2
google-crc32c==1.5.0
google-generativeai==0.5.4
google-pasta==0.2.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
graphviz==0.20.1
greenlet==3.0.3
groq==0.8.0
grpc-google-iam-v1==0.13.0
grpcio==1.64.0
grpcio-status==1.62.2
h11==0.14.0
h5py==3.8.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.23.4
hyperlink==21.0.0
hypothesis==6.108.2
idna==3.4
imageio==2.34.1
imageio-ffmpeg==0.5.1
incremental==22.10.0
iniconfig==2.0.0
install==1.3.5
intel-openmp==2021.4.0
interface-meta==1.3.0
ipykernel==6.23.2
ipython==7.34.0
ipywidgets==8.1.3
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.4
joblib==1.2.0
json5==0.9.25
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==7.4.9
jupyter_core==5.3.0
jupyter_server==2.14.1
jupyter_server_terminals==0.5.3
jupyterlab==4.2.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
jupyterlab_widgets==3.0.11
keras==2.11.0
kiwisolver==1.4.4
langchain==0.2.1
langchain-community==0.2.1
langchain-core==0.2.3
langchain-text-splitters==0.2.0
langsmith==0.1.67
lazy_loader==0.4
libclang==15.0.6.1
librosa==0.10.2.post1
lifelines==0.29.0
lightgbm==3.3.5
llvmlite==0.42.0
Markdown==3.4.1
MarkupSafe==2.1.2
marshmallow==3.21.2
matplotlib==3.6.3
matplotlib-inline==0.1.6
MiniSom==2.3.2
mistune==3.0.2
mkl==2021.4.0
mlxtend==0.23.1
mock==5.1.0
moviepy==1.0.3
mplfinance==0.12.10b0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
mypy-extensions==1.0.0
mysql==0.0.3
mysql-connector-python==9.0.0
mysqlclient==2.2.4
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.5.6
networkx==3.3
nltk==3.8.1
notebook==7.2.1
notebook_shim==0.2.4
numba==0.59.1
numpy==1.24.2
oauthlib==3.2.2
opencv-python==4.10.0.82
openpyxl==3.1.0
opt-einsum==3.3.0
Optunity==1.0.1
orjson==3.10.3
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.2
pandas==2.2.2
pandas-flavor==0.6.0
pandocfilters==1.5.1
parso==0.8.3
patsy==0.5.6
pi==0.1.2
pickleshare==0.7.5
Pillow==9.4.0
pingouin==0.5.4
platformdirs==3.0.0
plotly==5.13.0
pluggy==1.5.0
pooch==1.8.1
portpicker==1.6.0
proglog==0.1.10
prometheus_client==0.20.0
prompt-toolkit==3.0.38
proto-plus==1.23.0
protobuf==3.20.3
psutil==5.9.5
pSyCH==1.0.0
psycopg2==2.9.9
pyarrow==17.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.22
pydantic==2.7.2
pydantic_core==2.18.3
pydot==1.4.2
pydot-ng==2.0.0
pydotplus==2.0.2
pydu==0.7.2
pydub==0.25.1
pyelm==0.1.0
Pygments==2.15.1
PyJWT==2.8.0
pyOpenSSL==24.2.1
pyparsing==3.0.9
pypdf==4.2.0
pypng==0.20220715.0
PySocks==1.7.1
pytest==8.3.1
pytest-cov==5.0.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2022.7.1
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==24.0.1
qrcode==7.4.2
qtconsole==5.5.2
QtPy==2.4.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.1
rsa==4.9
safetensors==0.4.3
scikit-fuzzy==0.4.2
scikit-learn==1.2.1
scipy==1.10.0
seaborn==0.13.2
selenium==4.21.0
Send2Trash==1.8.3
shapely==2.0.4
simplejson==3.19.2
six==1.16.0
sklearn==0.0.post1
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
SpeechRecognition==3.10.4
spyder-kernels==2.3.3
SQLAlchemy==2.0.30
sqlparse==0.5.0
squarify==0.4.3
statsmodels==0.14.2
stemgraphic==0.9.1
sympy==1.12.1
tabpy==2.10.0
tabulate==0.9.0
tbb==2021.13.0
tenacity==8.2.1
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.30.0
termcolor==2.2.0
terminado==0.18.1
textblob==0.18.0.post0
threadpoolctl==3.1.0
tinycss2==1.3.0
tokenizers==0.19.1
tomli==2.0.1
toolz==0.12.1
torch==2.3.1
tornado==6.3.2
tqdm==4.66.4
traitlets==5.9.0
transformers==4.42.3
trio==0.25.1
trio-websocket==0.11.1
Twisted==24.3.0
twisted-iocpsupport==1.0.4
types-python-dateutil==2.9.0.20240316
typing-inspect==0.9.0
typing_extensions==4.12.0
tzdata==2024.1
uri-template==1.3.0
uritemplate==4.1.1
urllib3==1.26.14
vertexai==1.49.0
virtualenv==20.19.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==2.2.2
widgetsnbextension==4.0.11
wrapt==1.14.1
wsproto==1.2.0
xarray==2024.6.0
xgboost==2.1.0
yarl==1.9.4
zope.interface==6.4.post2
